refactor(actions): fold the two real red-team assertions into compound.test.ts, drop src/security/
Both files in src/security/ re-tested paths already covered elsewhere,
sometimes more weakly. Every pairing was checked against its counterpart
before deleting: the not-primitive, grant-stripping, risk-understating
and disabled-step-tool cases against compound.test.ts (whose versions
also assert the error message and the full ctx pass-through); the
away/actAs cases against registry.test.ts and fixture.e2e.test.ts (whose
versions drive a real HTTP server and additionally assert the actAs
tag); the unannotated-connector-risk case against connectors.test.ts.

Two assertions were genuinely unique and are now folded into
compound.test.ts's own seam tests: the walker never touches the
registry's own fetch when steps succeed, and a compound with no
invokeTool leaves a connector's execute untouched as well as the host
fetch.

Also: loadoutSeed's filter could never drop anything (a descriptor is
pushed and a dispatch entry set under the same condition, keyed by the
same name), and validateCapabilities, CapabilityIssue and
PrimitiveStepTarget were re-exported from the package root with no
consumer outside the compound walker.
