+++ packages/actions/src/index.ts
@@ -16,7 +16,6 @@ export { mcpConnector, type McpAuthContext, type McpHeadersResolver } from "./co
 export { createActions, type ActionsRegistry, type ActionsRunContext, type ServerActionHandler } from "./runtime/registry.js";
 export { createConnectGate, type ConnectGate, type ConnectGateOptions } from "./runtime/connect-gate.js";
 export { type ToolSearchMatch, type ToolSearchOptions } from "./runtime/search.js";
-export { validateCapabilities, type CapabilityIssue, type PrimitiveStepTarget } from "./runtime/compound.js";
 // Build-/dev-time extraction surface moved to `@vendoai/actions/sync` so the
 // runtime entry stays portable (no node:fs / TypeScript compiler in server
 // bundles). See src/sync/public.ts.
